+++ MercadoPagoSDK/MercadoPagoSDK/Builders/PXPaymentTypeChargeRuleBuilder.swift
@@ -1,0 +1,70 @@
+//
+//  PXPaymentTypeChargeRuleBuilder.swift
+//  MercadoPagoSDKV4
+//
+//  Created by Jonathan Scaramal on 21/09/2021.
+//
+
+import Foundation
+
+public class PXPaymentTypeChargeRuleBuilder {
+    
+    private var paymentTypeId: String
+    private var amount: Double
+    
+    public private(set) var taxable: Bool = true
+    public private(set) var label: String? = nil
+    public private(set) var detailModal: UIViewController? = nil
+    
+    public init(paymentTypeId: String, amount: Double) throws {
+        if amount == 0.0 {
+            throw PXPaymentTypeChargeRuleBuilderError.invalidAmount
+        }
+        
+        self.paymentTypeId = paymentTypeId
+        self.amount = amount
+    }
+    
+    @discardableResult
+    public func setTaxable(_ taxable: Bool) -> PXPaymentTypeChargeRuleBuilder {
+        self.taxable = taxable
+        return self
+    }
+    
+    @discardableResult
+    public func setLabel(_ label: String) -> PXPaymentTypeChargeRuleBuilder  {
+        self.label = label
+        return self
+    }
+    
+    @discardableResult
+    public func setDetailModal(_ detailModal: UIViewController) -> PXPaymentTypeChargeRuleBuilder  {
+        self.detailModal = detailModal
+        return self
+    }
+    
+    public func build() -> PXPaymentTypeChargeRule {
+        var chargeRule = PXPaymentTypeChargeRule(paymentTypeId: paymentTypeId, amountCharge: amount, detailModal: detailModal)
+        
+        chargeRule.label = label
+        chargeRule.taxable = taxable
+        
+        return chargeRule
+    }
+}
+
+public enum PXPaymentTypeChargeRuleBuilderError: Error {
+    case invalidAmount
+}
+
+extension PXPaymentTypeChargeRuleBuilderError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case .invalidAmount:
+            return NSLocalizedString(
+                "Charge rule amount cannot be zero",
+                comment: ""
+            )
+        }
+    }
+}