+++ MercadoPagoSDK/MercadoPagoSDK/LaytoutExtension/AnchorType.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+enum AnchorType {
+    case top
+    case bottom
+    case trailing
+    case leading
+    case centerX
+    case centerY
+    case width
+    case height
+}